Improve test runner behavior when client times out, crashes, or exits prematurely (#844)
While troubleshooting some other test cases, and trying to figure out
which test cases were causing a client impl to hang, I noticed issues
with the output of the test runner when this sort of thing happens:

1. If `-v` or `--vv` is used, it can be seen that the runner will print
that it's trying to run more test cases and send one or more tests to
the client, even though it has failed (timed out usually, but can also
happen if it crashes). So I've fixed it up so that once the client fails
or times out, we don't try to do anything else and don't print anything
else about trying to run other test cases.
2. If `--vv` is used, the pending test cases (at least one of which
would be the culprit for the client hanging or crashing) all get printed
in "Received response for ..." messages, even though no response was
actually received from the client.
3. When the client timed out or crashed, the process would abort and not
bother showing any other results. This can be kind of a pain -- it is
often useful to see the results that _were_ computed for all of the
tests that completed prior to the client hanging or crashing.
4. Finally, unrelated to clients hanging/crashing: the `-v` output that
printed the number of server configurations was misleading. For example,
when running `make runconformance` in this repo, it prints 19
configurations for the reference client and server. _But_ it then
proceeds in printing "Running X tests with ..." 20 or 22 times. This is
because it actually starts different processes when testing with the
gRPC implementations. So there may be 19 distinct configurations like
`{HTTP_VERSION_2, PROTOCOL_GRPC, TLS:false}`, there actually 20 or 22
servers started, some being used with the gRPC reference implementation.

All of these issues are resolved in this commit.