Must percent-encode gRPC message (#847)
I thought I had tested the change in #840 against a bunch of other
implementations, but I clearly missed this. In that PR, the code was
not correctly percent-encoding the "grpc-message" trailer. It turns out
that most implementations (including grpc-java, grpc-go, and connect-go)
were all lenient in the face of this issue, which is _how_ I missed
this.

What was happening is that `net/http` was _dropping_
the invalid header (since one of the tests has a newline in the message,
which is not allowed in headers) but it was still being picked up from
the `grpc-status-details-bin` header. Connect-kotlin seems to be
the only implementation that insists on using the `grpc-message`
value as the error message, regardless of what's in
`grpc-status-details-bin`, so a test case failed since the error message
it reported did not match the expected value. (I could have _sworn_
I tested that commit against kotlin last week, but I guess I didn't
actually do that correctly.)

So this fixes the issue: when the reference server synthesizes a raw
response, to reply with independent header and trailer metadata, it now
correctly encodes the "grpc-message" key.

The _bulk_ of this PR is actually adding new checks in the reference
client that could have caught this issue: it examines the "grpc-status",
"grpc-message", and "grpc-status-details-bin" values to verify that
the server implementation adheres to the specs. It also applies
similar checks to other binary metadata.

There are a handful of small other changes in here too that I addressed
while troubleshooting and fixing this issue.