Fix reference server issues related to CORS (#855)
The reference server was already configured to use CORS and to be
permissive. But there were some quirks to its config that could be
improved (especially if anyone were to test using TLS client certs,
which means we have to allow credentials in the cross-origin requests).

Also, the "raw response" capability of the reference server was
interfering with the CORS middleware. So this fixes that, which in turn
activates some alternate behavior in the gRPC Web JS client: apparently,
some of the behavior observed was actually the result of failed
CORS access control checks and not the intended behavior of the
library 🤦.

On a positive note: it is no longer necessary/useful to add access
control headers to raw responses. The CORS middleware now handles that
correctly.