When reference server is in HTTP 1.1 mode, don't allow negotiating HTTP/2 via TLS (#856)
Some clients may not have the direct ability to _disable_ negotiating
HTTP/2 if the server advertises its support during ALPN step of TLS
handshake. So to prevent such clients from using HTTP/2, if HTTP 1.1. is
the intent, we disable HTTP/2 in the reference server. (Sadly, we can't
also disable HTTP 1.1 when HTTP/2 is the intent since Go's `net/http`
doesn't provide a way to do that 🤷.)