Update test cases for consistency, backfill missing cancel tests (#874)
After noticing some naming inconsistencies, I added a test that verifies
that the embedded test files all conform to the naming conventions
described in the docs.

That found a few issues, so this PR fixes them all.

I noticed and changed a few other things while doing this other clean-up
in the tests:
1. We were being lenient with gRPC implementations regarding an unknown
content-type. That's because I thought this behavior (using a 415 HTTP
status code) wasn't fully specified. But it turns out that is is!
So I've moved these all to a test suite that is not protocol-specific,
since all three protocols now expect the same behavior.
2. I noticed a test that was commented out with a TODO. The change the
TODO was contingent was on was already done, so the test could be
re-enabled.
3. The new naming convention test revealed a naming issue in the
cancellation tests. When I went to fix that, I also noticed there were
some redundant cases in here: the `cancel-after-zero-responses` test
cases are not actually different than the `cancel-after-close-send`.
While the contents of the test case are slightly different, the behavior
they induce in the client is the same, which is to cancel the RPC right
after the request stream is closed. It seemed odd that we'd have a test
for cancelling "after" receiving zero responses. That's really like a
cancel before the first response is received, and that felt awkward to
implement since ideally you simply check that number _after_ receiving a
response (per the name of the field). So I opted to remove those and
keep the `cancel-after-close-send` version.
4. The server-stream cancel tests only had the
`cancel-after-zero-responses`, and no `cancel-after-close-send` version.
So I "fixed" that -- to use `cancel-after-send` for consistency with the
others -- only to realize that the reference clients did not actually
implement it. That's when I also realized that there was no test for
cancelling unary RPCs. (The reference clients also did not implement
cancellation in the unary RPC flow.)

The biggest changes in here were to get all of the client
implementations to handle the `cancel.after_close_send_ms` field in the
input.