Revise some of the docs, add information about upgrading (#876)
I was reviewing some of the docs and noticed a few things that could be
improved.

The changes to the bottom section of the README stem from
thinking about other kinds of changes that we must reasonably make from
time to time that could prevent a user from easily being able to upgrade
to a newer version of the suite. I figured being very explicit about
this in this document is best for clarity and transparency.

This also adds a section to the README that describes the machinery of
the thing and also includes a sequence diagram to show what's going on
inside the test runner.

And it adds a section in other docs that describe the process for upgrading
to a new release of the conformance suite.