Require that client implementations properly enforce timeouts (#877)
Because the connect-go server framework will try to curtail work beyond
the deadline, the timeout test cases weren't working quite as expected.
For the streaming endpoints, the deadline would elapse and cause the
server handler's stream interactions to return a `deadline_exceeded`
error, which it would then return to the client. So the client was
correctly reporting a `deadline_exceeded` error for the streaming
tests, even when it didn't actually enforce a deadline.

So this change causes the reference server to remove the timeout
headers, so that they won't result in a context deadline being created
for the handler. That way, the only way the tests can result in a
`deadline_exceeded` error is if the client is actually enforcing it.