Add conformance test cases for handling empty (zero byte) messages (#918)
Adds new test cases to make sure that servers correctly handle requests
of zero bytes and clients handle responses of zero bytes. The responses
required the use of "raw responses" in the test case, because
conformance server implementations by default never send an empty
response message, since they echo back the request information in the
response.

This also does some tweaking of the test cases in `basic.yaml` to remove
redundant/unnecessary test cases. I also tried adding "no-requests" test
cases, to go along with existing "no responses" test cases. However,
this uncovered an issue between the reference client and the grpc server
that is still eluding me. So I've left those new test cases commented
out for now and can uncomment this after I have time to debug them a
little more.

Finally, this updates all of the client implementations to add an empty
`ConformancePayload` message for a response, even if that response did
not contain a `payload` field. This is specifically for handling these
empty responses. This also was worth an update to the docs so we provide
guidance to users that are implementing clients on how to handle these.