Test with go 1.22 and 1.23 (#946)
The minimum for this repo is now 1.22, not 1.21, due to upgrades in both
grpc-go and quic-go. In other projects, we support the past three versions
of Go, instead of just past two. But this repo isn't a library for our users
(some of which may use Go 1.21) and isn't a tool that our users need to
build from source (possibly using Go 1.21). Instead, we provide release
binaries that users should download if they are unable to build from
source. So dropping 1.21 should be fine for this repo.

Much of the diff is due to a handful of new lint checks -- updating
golangci-lint was required in order to support 1.23, and it includes
several new checks, including one that wants string constants/literals
used to query HTTP headers to use canonical case (because it is more
efficient: not using canonical case forces the `http.Header` value to
build a new canonicalized string before querying the underlying map).

Also, moving to 1.22 means the default loop-variable behavior no longer
captures the address, so we no longer need to make a copy (and there's a
linter that caught places we were doing this, so they've all been updated).