Allow server to cancel when deadline expires (#999)
This makes some assertions for timeout cases a little more lenient. The
tests previously asserted that a server would return a "deadline
exceeded" error when the client's deadline expires. However, not all
gRPC servers behave this way. The gRPC Go reference implementation was
recently updated to sometimes just cancel the stream instead of sending
back an error response. This doesn't seem unreasonable, so instead of
updating the "known failing" list of cases for the grpc-go server, this
PR makes this behavior allowed (which means it will be allowed in any
other implementation-under-test, too).

The test case structure lets us define other acceptable error codes for
test cases, however we don't actually want to add "canceled" as
generally acceptable for these tests. It's really only acceptable for
the server (clients should always report "deadline exceeded" for these
cases). So we need to update the test assertions logic to get that level
of nuance.